+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#115](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/115)
 * Populating local DB from production via pgsync
   [#118](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/118)
+* Meta table for tracking applied migrations
+  [#119](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/119)
 
 ### Fixed